Rename test package
