+++ pkg/services/authn/clients/ldap_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/grafana/grafana/pkg/services/ldap/multildap"
 	"github.com/grafana/grafana/pkg/services/ldap/service"
 	"github.com/grafana/grafana/pkg/services/login"
-	"github.com/grafana/grafana/pkg/services/login/logintest"
+	"github.com/grafana/grafana/pkg/services/login/authinfotest"
 	"github.com/grafana/grafana/pkg/services/org"
 	"github.com/grafana/grafana/pkg/services/user"
 	"github.com/grafana/grafana/pkg/services/user/usertest"
@@ -192,7 +192,7 @@ func setupLDAPTestCase(tt *ldapTestCase) *LDAP {
 			return nil
 		},
 	}
-	authInfoService := &logintest.AuthInfoServiceFake{
+	authInfoService := &authinfotest.AuthInfoServiceFake{
 		ExpectedUserAuth: &tt.expectedAuthInfo,
 		ExpectedError:    tt.expectedAuthInfoErr,
 	}
