+++ pkg/services/login/authinfotest/logintest.go
@@ -1,4 +1,4 @@
-package logintest
+package authinfotest
 
 import (
 	"context"
